separating react components into different files
