+++ app.js
@@ -1,5 +1,3 @@
-import Hello from './app/components/hello.jsx';
-import World from './app/components/world.jsx';
-import CommentBox from './app/components/CommentBox.jsx';
-
 require("./app/assets/css/my-style.css");
+
+import CommentBox from './app/components/CommentBox.jsx';
